% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bindea2013c_enrich}
\alias{Bindea2013c_enrich}
\title{Cell Interaction Enrichment from function TimiEnrich 
with cell type meaker annotated by Bindea2013_cancer}
\format{
a data frame of enrichment result:
\describe{
  \item{Index}{A numeric vector used to specify which cell pair would be used in visualization}
  \item{Rank}{A rank of cell interaction according to Adjust. P-value}
  \item{Cell.Pair}{Cell Interaction, e.g. Cell A_Cell B}
  \item{Favorable.Cell.Type}{In the cell pair A_B, the first cell A}
  \item{Unfavorable.Cell.Type}{In the cell pair A_B, the second cell B}
  \item{No.Total.Marker.Pair}{The total number of marker pairs in the cell pair annotation}
  \item{No.Shared.Marker.Pair}{The number of marker pairs shared by query pairs and annotation}
  \item{Enrichment.Ratio}{Enrichment Ratio}
  \item{P.Value}{P. value}
  \item{Adjust.P.Value}{Adjust.P.Value}
  \item{Shared.Marker.Pair}{The marker pairs shared by query pairs and annotation}
  \item{Total.Marker.Pair}{The marker pairs in the annotation}
}
}
\source{
example result generated from function TimiEnrich(example/example01_Bindea2013_Cancer.R)
}
\usage{
data(Bindea2013c_enrich)
}
\description{
An example result of cell interaction enrichment generated from function TimiEnrich
}
\keyword{example}
\keyword{result}
