% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellType_Alex2020_Levi2019_TME}
\alias{CellType_Alex2020_Levi2019_TME}
\title{Cell Type in melanoma microenvironment generated from Alex2020 and Levi2019 with cancer cells}
\format{
A data frame with 663 rows and 3 variables:
\describe{
  \item{CellType}{Immune cell type, skin cell type and cancer cell type}
  \item{Gene}{Marker gene of the cell type}
  \item{Dataset}{The source of the annotation}
}
}
\source{
Modification can be found in inst/extdata/build_CellType_Alex2020_Levi2019_TME.R
}
\usage{
data(CellType_Alex2020_Levi2019_TME)
}
\description{
A dataset containing the cell type and markers annotated by Alex2020 and Levi2019
}
\references{
Hughes, T. K., Wadsworth II, M. H., Gierahn, T. M., Do, T., Weiss, D., Andrade, P. R., ... & Shalek, A. K. (2020). Second-strand synthesis-based massively parallel scRNA-seq reveals cellular states and molecular features of human inflammatory skin pathologies. Immunity, 53(4), 878-894.

Tirosh, I., Izar, B., Prakadan, S. M., Wadsworth, M. H., Treacy, D., Trombetta, J. J., ... & Garraway, L. A. (2016). Dissecting the multicellular ecosystem of metastatic melanoma by single-cell RNA-seq. Science, 352(6282), 189-196.
}
\keyword{datasets}
