% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellType_Bindea2013_cancer}
\alias{CellType_Bindea2013_cancer}
\title{Cell Type generated from Bindea2013 with cancer cells}
\format{
A data frame with 545 rows and 3 variables:
\describe{
  \item{CellType}{Immune cell type and cancer cell type}
  \item{Gene}{Marker gene of the cell type}
  \item{Dataset}{The source of the annotation}
}
}
\source{
Modification can be found in inst/extdata/build_CellType_Bindea2013_cancer.R
}
\usage{
data(CellType_Bindea2013_cancer)
}
\description{
A dataset containing the cell type and markers annotated by Bindea2013
}
\references{
Bindea, G., Mlecnik, B., Tosolini, M., Kirilovsky, A., Waldner, M., 
Obenauf, A. C., ... & Galon, J. (2013). Spatiotemporal dynamics of intratumoral 
immune cells reveal the immune landscape in human cancer. Immunity, 39(4), 782-795.
}
\keyword{datasets}
