% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellType_Charoentong2017_Galon2013_TIP_Immune}
\alias{CellType_Charoentong2017_Galon2013_TIP_Immune}
\title{Immune Cell Type generated from Charoentong2017, Galon2013, TIP}
\format{
A data frame with 1409 rows and 3 variables:
\describe{
  \item{CellType}{Immune cell type}
  \item{Gene}{Marker gene of the cell type}
  \item{Dataset}{The source of the annotation}
}
}
\source{
Modification can be found in inst/extdata/build_CellType_Charoentong2017_Galon2013_TIP_Immune.R
}
\usage{
data(CellType_Charoentong2017_Galon2013_TIP_Immune)
}
\description{
A dataset containing the cell type and markers annotated by Charoentong2017, Galon2013, TIP
}
\references{
Charoentong, P., Finotello, F., Angelova, M., Mayer, C., Efremova, M., Rieder, D., ... & Trajanoski, Z. (2017). Pan-cancer immunogenomic analyses reveal genotype-immunophenotype relationships and predictors of response to checkpoint blockade. Cell reports, 18(1), 248-262.

Bindea, G., Mlecnik, B., Tosolini, M., Kirilovsky, A., Waldner, M., Obenauf, A. C., ... & Galon, J. (2013). Spatiotemporal dynamics of intratumoral immune cells reveal the immune landscape in human cancer. Immunity, 39(4), 782-795.

Xu, L., Deng, C., Pang, B., Zhang, X., Liu, W., Liao, G., ... & Li, X. (2018). TIP: a web server for resolving tumor immunophenotype profiling. Cancer research, 78(23), 6575-6580.
}
\keyword{datasets}
