% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellType_Newman2015_LM22}
\alias{CellType_Newman2015_LM22}
\title{Immune Cell Type generated from LM22}
\format{
A data frame with 1140 rows and 3 variables:
\describe{
  \item{CellType}{Immune cell type}
  \item{Gene}{Marker gene of the cell type}
  \item{Dataset}{The source of the annotation}
}
}
\source{
Modification can be found in inst/extdata/build_CellType_Newman2015_LM22.R
}
\usage{
data(CellType_Newman2015_LM22)
}
\description{
A dataset containing the cell type and markers annotated by LM22, 
which is the cell-type signature of CIBERSORT and CIBERSORTx.
}
\references{
Newman, A. M., Liu, C. L., Green, M. R., Gentles, A. J., Feng, W., Xu, Y., ... & Alizadeh, A. A. (2015). Robust enumeration of cell subsets from tissue expression profiles. Nature methods, 12(5), 453-457.
}
\keyword{datasets}
