% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellType_Tirosh2016_melanoma_TME}
\alias{CellType_Tirosh2016_melanoma_TME}
\title{Cell Type in melanoma microenvironment generated from Hughes2020 and Tirosh2016 with cancer cells}
\format{
A data frame with 391 rows and 3 variables:
\describe{
  \item{CellType}{Immune cell type, skin cell type and cancer cell type}
  \item{Gene}{Marker gene of the cell type}
  \item{Dataset}{The source of the annotation}
}
}
\source{
Modification can be found in inst/extdata/build_CellType_Tirosh2016_melanoma_TME.R
}
\usage{
data(CellType_Tirosh2016_melanoma_TME)
}
\description{
A dataset containing the cell type and markers annotated by Tirosh2016
}
\references{
Tirosh, I., Izar, B., Prakadan, S. M., Wadsworth, M. H., Treacy, D., Trombetta, J. J., ... & Garraway, L. A. (2016). Dissecting the multicellular ecosystem of metastatic melanoma by single-cell RNA-seq. Science, 352(6282), 189-196.
}
\keyword{datasets}
