% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellType_TNBC_aPDL1}
\alias{CellType_TNBC_aPDL1}
\title{Cell Type customized for TNBC (anti-PD1/PD-L1)}
\format{
A data frame with 449 rows and 3 variables:
\describe{
  \item{CellType}{Immune cell type}
  \item{Gene}{Marker gene of the cell type}
  \item{Dataset}{The source of the annotation}
}
}
\source{
For more details, please refer: https://github.com/CSkylarL/MSofTimiGP-Response/blob/master/Fig1/code/Fig1.script2_generate_TimiGP_markers.R
}
\usage{
data(CellType_TNBC_aPDL1)
}
\description{
A dataset containing the cell type and markers customized for 
Triple-negative breast cancer (TNBC) cohorts 
treated with anti-PD1/PD-L1 therapy. 
This cell type annotation generated from single cell RNA-seq (GSE169246) 
and it includes 11 immune cell types consistent with 
the imaging mass cytometry (IMC) (zenodo.7990870).
}
\references{
Li C1,, HongW, Reuben A,Wang L, Anirban M, Zhang J, Cheng C. TimiGP-Response Module: Illustrates the pan-cancer immune landscape associated with response to immunotherapy. bioRxiv. (2024)
}
\keyword{datasets}
