% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellType_Zheng2021_Tcell}
\alias{CellType_Zheng2021_Tcell}
\title{T Cell Subtypes generated from Zheng2021}
\format{
A data frame with 926 rows and 3 variables:
\describe{
  \item{CellType}{T cell subtype}
  \item{Gene}{Marker gene of the cell type}
  \item{Dataset}{The source of the annotation}
}
}
\source{
Modification can be found in inst/extdata/build_CellType_Zheng2021_Tcell.R
}
\usage{
data(CellType_Zheng2021_Tcell)
}
\description{
A dataset containing the cell type and markers annotated by Zheng2021
}
\references{
Zheng, L., Qin, S., Si, W., Wang, A., Xing, B., Gao, R., ... & Zhang, Z. (2021). Pan-cancer single-cell landscape of tumor-infiltrating T cells. Science, 374(6574), abe6474.
}
\keyword{datasets}
