% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Immune_Marker_n1326}
\alias{Immune_Marker_n1326}
\title{Immune markers generated from Charoentong2017, Galon2013, TIP and classic markers}
\format{
A vector of 1326 immune markers
}
\source{
Modification can be found in inst/extdata/build_Immune_Marker_n1326.R
}
\usage{
data(Immune_Marker_n1326)
}
\description{
A dataset containing the immune markers annotated by Charoentong2017, Galon2013, TIP 
and additional classic immune inhibitors, stimulators and checkpoints.
}
\references{
Charoentong, P., Finotello, F., Angelova, M., Mayer, C., Efremova, M., Rieder, D., ... & Trajanoski, Z. (2017). Pan-cancer immunogenomic analyses reveal genotype-immunophenotype relationships and predictors of response to checkpoint blockade. Cell reports, 18(1), 248-262.

Bindea, G., Mlecnik, B., Tosolini, M., Kirilovsky, A., Waldner, M., Obenauf, A. C., ... & Galon, J. (2013). Spatiotemporal dynamics of intratumoral immune cells reveal the immune landscape in human cancer. Immunity, 39(4), 782-795.

Xu, L., Deng, C., Pang, B., Zhang, X., Liu, W., Liao, G., ... & Li, X. (2018). TIP: a web server for resolving tumor immunophenotype profiling. Cancer research, 78(23), 6575-6580.
}
\keyword{datasets}
