% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Newman2015_COX_MP_SKCM06}
\alias{Newman2015_COX_MP_SKCM06}
\title{COX regression Results from function TimiCOX with cell type meaker annotated by Newman2015_LM22}
\format{
A data frame with 115440 rows and 3 variables:
\describe{
  \item{Row name}{Marker pair}
  \item{HR}{Hazard.Ratio}
  \item{PV}{P-Value}
  \item{QV}{Adjust P-value}
}
}
\source{
intermediate result generated from function TimiCOX
}
\usage{
data(Newman2015_COX_MP_SKCM06)
}
\description{
An intermediate result generated from function TimiCOX
that reveals the association between each marker pairs and favorable prognosis.
}
\keyword{intermediate}
\keyword{result}
