% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SKCM06info}
\alias{SKCM06info}
\title{TCGA metastatic melanoma(SKCM06) clinical infomation}
\format{
A data frame with 368 rows and 2 variables:
\describe{
  \item{row}{Patient ID}
  \item{column}{Clinical information, including event(death) and time(day) to event}
}
}
\source{
Modification can be found in inst/extdata/build_SKCM06_RNA_info.R
}
\usage{
data(SKCM06info)
}
\description{
A dataset containing 20,501 genes from 368  metastatic melanoma samples.
}
\references{
\url{"https://gdac.broadinstitute.org/"}
}
\keyword{datasets}
