% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SKCM06rna}
\alias{SKCM06rna}
\title{TCGA metastatic melanoma(SKCM06) RNA expression}
\format{
A data frame with 20501 rows and 368 variables:
\describe{
  \item{row}{Gene}
  \item{column}{Sample(patient) ID}
}
}
\source{
Modification can be found in inst/extdata/build_SKCM06_RNA_info.R
}
\usage{
data(SKCM06rna)
}
\description{
A dataset containing 20,501 genes from 368  metastatic melanoma samples.
}
\references{
\url{"https://gdac.broadinstitute.org/"}
}
\keyword{datasets}
