% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich.R
\name{TimiBG}
\alias{TimiBG}
\title{Generate marker pair background}
\usage{
TimiBG(marker.pair = NULL)
}
\arguments{
\item{marker.pair}{a vector of marker pairs connected by "_"}
}
\value{
A vector of background gene pairs
}
\description{
Given a gene pair A_B, it will generate 
the background gene pairs A_B and B_A
}
\examples{
\dontrun{
  data(Galon2013c_COX_MP_SKCM06)
  cox_res <- Galon2013c_COX_MP_SKCM06
  background <- TimiBG(marker.pair = row.names(cox_res))
}
}
\author{
Chenyang Skylar Li
}
