% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChordDiagram.R
\name{TimiCellChord}
\alias{TimiCellChord}
\title{Chord diagram of Cell Interaction}
\usage{
TimiCellChord(
  resdata = NULL,
  select = NULL,
  dataset = NULL,
  group = NULL,
  color = NULL
)
}
\arguments{
\item{resdata}{TimiGP enrichment result generated from TimiEnrich}

\item{select}{a numeric vector of selected cell pairs according to "Index" column in resdata. 
Default selection is all statistically significantly enriched cell pairs(adjusted p value < 0.05).}

\item{dataset}{a value in one of 
c("Galon2013","Galon2013_Cancer","Charoentong2017", 
"Alex2020_Levi2019","Zhang2021S2","TIP","Other"). 
The first four options include default group and color settings. 
If you use other dataset or want to change group and colors, 
please choose "Other".}

\item{group}{a vector of self-defined cell groups, 
whose names are the all favorable and unfavorable cells types in selection.}

\item{color}{a vector of self-defined cell colors, 
whose names are the all favorable and unfavorable cells types in selection.}
}
\value{
Chord diagram of Cell Interaction
}
\description{
Chord diagram reveals the cell interaction(cell pair) associated with good prognosis.
The arrow points from favorable cell type A to unfavorable cell type B, 
which denotes that high A-to-B ratio associated with a good prognosis.
The width of the arrow represents -log10(Adjust.P.Value),namely,
the wider the arrow is, the smaller the adjusted p-value is.
}
\examples{
\dontrun{
  data("Galon2013c_enrich")
  res <- Galon2013c_enrich
  TimiCellChord(resdata = res,dataset = "Galon2013_Cancer")
  TimiCellChord(resdata = res,dataset = "Galon2013_Cancer",select = 1:10)
 
}
}
\author{
Chenyang Skylar Li
}
