% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pairwise.R
\name{TimiCellPair}
\alias{TimiCellPair}
\title{Generate cell pair representing potential interaction using 
the corresponding marker pair annotation}
\usage{
TimiCellPair(geneset = NULL, dataset = NULL, core = 1)
}
\arguments{
\item{geneset}{a data.frame of cell markers, 
in which the 1st column is cell type, 
the 2nd column is the marker gene 
and the 3rd column is the name of the dataset (optional)}

\item{dataset}{specified at least 1 dataset occured in the 3rd column of geneset}

\item{core}{a numeric value shows the number of cores 
to use for parallel execution. The default value is 1.}
}
\value{
A dataframe of Cell pair and the corresponding marker pair
}
\description{
Given any two different cell types,
Cell A has markers a1 and a2 and cell B has markers b1 and b2.
Cell pair A_B includes marker pairs: a1_b1, a1_b2, a2_b1, a2_b2;
Cell pair B_A includes marker pairs: b1_a1, b1_a2, b2_a1, b2_a2
}
\examples{
\dontrun{
  data(CellType_Galon2013_cancer)
  geneset <- CellType_Galon2013_cancer
  cell_pair <- TimiCellPair(geneset = geneset,core = 2)
}
}
\author{
Chenyang Skylar Li
}
