% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{TimiCheckEvent}
\alias{TimiCheckEvent}
\title{Check Clinical Info}
\usage{
TimiCheckEvent(info = NULL)
}
\arguments{
\item{info}{a data.frame in which the 1st column is event 
and the 2nd column is days-to-event.}
}
\value{
filtered clinical info
}
\description{
Check Event and Time-to-event; remove NAs and time-to-event <=0
}
\examples{
\dontrun{
  data(SKCM06info)
  dim(SKCM06info)
  info <- TimiCheckEvent(SKCM06info)
  dim(info) 
}
}
\author{
Chenyang Skylar Li
}
