% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cox.R
\name{TimiCOX}
\alias{TimiCOX}
\title{Find marker pairs associated with favorable prognosis}
\usage{
TimiCOX(
  mps = NULL,
  info = NULL,
  p.adj = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")
)
}
\arguments{
\item{mps}{a matrix of Marker Pair Score}

\item{info}{a data.frame in which the 1st column is event 
and the 2nd column is days-to-event.}

\item{p.adj}{p.adjust.methods. 
One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}
}
\value{
A list of two results: 
1. Result of cox regression(cox_res)
(HR: Hazard.Ratio,PV: P-Value,QV: Adjust P-value). 
2. modified marker pair score(mps)
}
\description{
Perform univariate cox regression that fits each marker pair as a variable.
If a Pair A_B associated with poor prognosis(HR > 1), even not significant, 
it will be changed to B_A and reverse its value in the matrix of pair
}
\examples{
\dontrun{
  data("SKCM06info")
  data("SKCM06rna")
  data("Immune_Marker_n1293")
  info <- TimiCheckEvent(SKCM06info)
  rna <- TimiPrePropress(gene = Immune_Marker_n1293,rna = SKCM06rna,cohort = rownames(info))
  mps <- TimiGenePair(rna)
  dim(mps)
  # TimiCOX 
  res <- TimiCOX(mps = mps,info = info,p.adj = "BH")
  mps <- res[["mps"]] 
  cox_res <- res[["cox_res"]]
}
}
\author{
Chenyang Skylar Li
}
