% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich.R
\name{TimiEnrich}
\alias{TimiEnrich}
\title{Enrichment Analysis}
\usage{
TimiEnrich(
  gene = NULL,
  background = NULL,
  geneset = NULL,
  p.adj = NULL,
  core = 1,
  pair = TRUE
)
}
\arguments{
\item{gene}{a vector of gene symbol}

\item{background}{background genes or pairs}

\item{geneset}{a data frame, column 1 shows cell(pair), column 2 shows gene(pair)}

\item{p.adj}{p.adjust.methods. 
One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{core}{a numeric value shows the number of cores to use for parallel execution. 
The default value is 1.}

\item{pair}{a logical value. If TRUE(default), perform enrichment analysis in cell pair. 
If using the function for general cell type enrichment analysis, please choose FALSE.}
}
\value{
A dataframe of enrichment results showing the cell pair associated with favorable prognosis
}
\description{
It identify cell pair associated with favorable prognosis 
by performing over-representation(enrichment analysis)
}
\examples{
\dontrun{
  # Generate Cell Pair Annotation: TimiCellPair
  data(CellType_Galon2013_cancer)
  geneset <- CellType_Galon2013_cancer
  cell_pair <- TimiCellPair(geneset = geneset,core = 20)
  # Select marker pairs A_B=1 associated with good prognosis
  data(Galon2013c_COX_MP_SKCM06)
  cox_res <- Galon2013c_COX_MP_SKCM06
  GP <- rownames(cox_res)[which(cox_res$QV<0.05)]
  # generate background: TimiBG
  background <- TimiBG(marker.pair = row.names(cox_res))
  # Enrichment Analysis: TimiEnrich
  res <- TimiEnrich(gene = GP, background = background, 
                    geneset = cell_pair, p.adj = "BH",core=20)
  
}
}
\author{
Chenyang Skylar Li
}
