% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher.R
\name{TimiFisher}
\alias{TimiFisher}
\title{Find marker pairs associated with response to therapy}
\usage{
TimiFisher(mps = NULL, info = NULL, p.adj = "BH", parallel = FALSE, core = 1)
}
\arguments{
\item{mps}{a matrix of Marker Pair Score}

\item{info}{a data.frame in which the 1st column contains 
two groups on response to therapy, "Responder == 1" & "Non-Responder == 0" 
and row name is patient ID.}

\item{p.adj}{p.adjust.methods. 
One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"    
The default value is "BH".}

\item{parallel}{a logical value. If TRUE, enable parallel computing. 
The default value if FALSE}

\item{core}{a numeric value shows the number of cores for parallel execution. 
The default value is 1.}
}
\value{
Result of fisher test
(Rowname: Marker Pair,PV: P-Value,QV: Adjust P-value).
}
\description{
Perform fisher test (two sided) to look for marker pair associated with response to therapy.
}
\examples{
\dontrun{ 
# See example07 for more details
# Load the example data
data(TNBCaPD1info)
data(TNBCaPD1rna)
# Load the example cell type annotation
data("CellType_TNBC_aPDL1")
geneset <- CellType_TNBC_aPDL1
marker <- unique(geneset$Gene)
# Preprocess
info <- TNBCaPD1info
rna <- TimiPrePropress(marker = marker, cohort = rownames(info),
                       log = F, GMNorm = T, rna = TNBCaPD1rna)
# Generate marker pair score
mps <- TimiGenePair(rna)
# Use the function TimiFisher 
res <- TimiFisher(mps = mps,info = info,p.adj = "BH")
# Assign results
mps <- res$mps
fisher_res <- res$fisher_res
}
}
\author{
Chenyang Skylar Li
}
