% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Favorability.R
\name{TimiFS}
\alias{TimiFS}
\title{Calulate Favorability Score of cell types}
\usage{
TimiFS(resdata = NULL, cutoff = 0.05)
}
\arguments{
\item{resdata}{TimiGP enrichment result generated from TimiEnrich}

\item{cutoff}{A cutoff of adjusted P-value used to filter cell interactions. 
The default cutoff is 0.05.}
}
\value{
A matrix of Favorability Score 
which includes favorable score and unfavorable score.
Difference = Favorable.Score - Unfavorable.Score. 
If Difference greater than 0, the cell type is classified as F(avorable) cell. 
If Difference is negative, the cell type is classified as F(avorable) cell.
}
\description{
It calculates the favorability score of each cell type 
based on TimiGP cell interaction network.
In the network, 
the out-degree of cell A means how many times high  A-to-other cell ratio 
is associated with favorable prognosis.
The in-degree of cell A means how many times low A-to-other cell ratio 
is  associated with favorable prognosis.
The favorability score  includes favorable score and unfavorable score.
Favorable score: out-degree of the cell/sum of out-degree of all cell*100.
Unfavorable score: in-degree of the cell/sum of in-degree of all cell*100.
}
\examples{
\dontrun{
  data("Galon2013c_enrich")
  res <- Galon2013c_enrich
  score <- TimiFS(res)
  head(score)
}
}
\author{
Chenyang Skylar Li
}
