% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Favorability.R
\name{TimiFSBar}
\alias{TimiFSBar}
\title{Visualization of favorability score}
\usage{
TimiFSBar(score = NULL, select = NULL)
}
\arguments{
\item{score}{Favorability score calculated from TimiFS}

\item{select}{A numeric vector of selected cell interactions according to "Index" column in resdata. 
Default selection is all cell type.}
}
\value{
A barplot of Favorability Score
}
\description{
It generate a bar plot of the favorability score of cell type 
to evaluate its favorable(orange,positive) or unfavorable(Blue,negative) role in prognosis.
}
\examples{
\dontrun{
  data("Bindea2013c_enrich")
  res <- Bindea2013c_enrich
  score <- TimiFS(res)
  head(score)
  p <- TimiFSBar(score)
  p
}
}
\author{
Chenyang Skylar Li
}
