% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChordDiagram.R
\name{TimiGeneChord}
\alias{TimiGeneChord}
\title{Chord diagram of Gene Interaction}
\usage{
TimiGeneChord(resdata = NULL, select = 1, color = NULL)
}
\arguments{
\item{resdata}{TimiGP enrichment result generated from TimiEnrich}

\item{select}{A numeric value of selected cell interactions according to "Index" column. 
Default selection is the top 1 cell interaction.}

\item{color}{A vector of self-defined marker colors, 
whose names are the enriched markers
(not pair,you need to split the genes in Shared.IMGP column) 
in selected cell interaction.}
}
\value{
Chord diagram of Gene Interaction
}
\description{
Chord diagram reveals the gene interactions of a selected cell interaction.
The arrow points from enriched marker F of favorable cell type A 
to enriched marker U of unfavorable cell type B, 
which denotes that 
the expression of F greater than that of U is associated with a good prognosis.
}
\examples{
\dontrun{
  data("Bindea2013c_enrich")
  res <- Bindea2013c_enrich
  TimiGeneChord(resdata = res)
  TimiGeneChord(resdata = res,select = 2)
 
}
}
\author{
Chenyang Skylar Li
}
