% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network.R
\name{TimiGeneNetwork}
\alias{TimiGeneNetwork}
\title{Generate gene network files for Cytoscape}
\usage{
TimiGeneNetwork(
  resdata = NULL,
  select = NULL,
  dataset = NULL,
  geneset = NULL,
  export = TRUE,
  path = NULL
)
}
\arguments{
\item{resdata}{TimiGP cox result generated from TimiCOX}

\item{select}{A  character vector of selected gene pairs according to rownames in resdata. 
Default selection is all inter-cell interactions(adjusted p value < 0.05).}

\item{dataset}{A  value in one of 
c("Bindea2013_Cancer","Immune3", 
"Newman2015", 
"Tirosh2016","Zheng2021","Other")
The first four options include default group and color settings. 
If you use other dataset or want to change group and colors, 
please choose "Other".}

\item{geneset}{A  data.frame of cell markers, 
in which the 1st column is cell type, 
the 2nd column is the marker gene 
and the 3rd column is the name of the dataset(optional)}

\item{export}{A  logical value. If TRUE, it will generate network files 
for Cytoscape analysis. The default value is TRUE.}

\item{path}{A  directory to export the network files when "export = TRUE".}
}
\value{
A list of network required files(network,node,edge)
}
\description{
It generates three files that can be used to build network in Cytoscape:
1. network files: simple interaction file (gene_network.sif);
2. node attributes (gene_node.txt);
3. edge attributes (gene_edge.txt).
The function also returns a list of above files that can be modified in R.
}
\examples{
\dontrun{
  data(Bindea2013c_COX_MP_SKCM06)
  cox_res <- Bindea2013c_COX_MP_SKCM06
  NET <- TimiGeneNetwork(resdata = cox_res,dataset = "Bindea2013_Cancer",export=T,path="./")
}
}
\author{
Chenyang Skylar Li
}
