% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pairwise.R
\name{TimiGenePair}
\alias{TimiGenePair}
\title{Capture logical/continuous relation between marker pairs}
\usage{
TimiGenePair(rna = NULL, cont = FALSE)
}
\arguments{
\item{rna}{a data.frame of preprocessed transcriptomic profile}

\item{cont}{a logical value. If TRUE, capture the continuous relation. 
The default is FALSE}
}
\value{
a matrix of Marker Pair Score
}
\description{
Default: Capture logical relation of any two marker pairs, 
and generate a matrix of Marker Pair Score:
1 or TRUE = the expression of gene A > that of gene B, 
0 or FALSE = the expression of gene A < that of gene B. 
\cr
Optional:  Capture continuous relation of any two marker pairs, 
and generate a matrix of Marker Pair Score:
the expression of gene A - that of gene B,
}
\examples{
\dontrun{
  data("SKCM06info")
  data("SKCM06rna")
  data("Immune_Marker_n1293")
  dim(SKCM06info)
  info <- TimiCheckEvent(SKCM06info)
  dim(info)
  dim(SKCM06rna)
  rna <- TimiPrePropress(gene = Immune_Marker_n1293,rna = SKCM06rna,cohort = rownames(info))
  # TimiGenePair
  mps <- TimiGenePair(rna)
  dim(mps)
}
}
\author{
Chenyang Skylar Li
}
