% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutation.R
\name{TimiPermFDR}
\alias{TimiPermFDR}
\title{Permutation of marker genes to control FDR}
\usage{
TimiPermFDR(
  resdata = NULL,
  geneset = NULL,
  gene = NULL,
  background = NULL,
  niter = 100,
  core = 1
)
}
\arguments{
\item{resdata}{TimiGP enrichment result generated from TimiEnrich}

\item{geneset}{a data.frame of cell markers used in TimiCellPair,
in which the 1st column is cell type, 
the 2nd column is the marker gene 
and the 3rd column is the name of the dataset (optional)}

\item{gene}{a vector of gene symbol used in TimiEnrich}

\item{background}{background genes or pairs used in TimiEnrich}

\item{niter}{A numeric value used to determine the number of permutation. 
The default cutoff is 100.}

\item{core}{a numeric value shows the number of cores 
for parallel execution. The default value is 1.}
}
\value{
A new column "Permutation.FDR" is appended to 
dataframe of inter-cell interactions generated from TimiEnrich
}
\description{
It conservatively permutates marker gene sets for all cell types 
by shuffling their marker genes while preserving the sizes of their sets. 
This permutation breaks the underlying association 
between the genes and the cell types, 
which enables the examination of cell-type marker quality 
and the control of false positives in identifying cell interactions. 
It exports the FDR (Q value) of each interaction.
}
\examples{
\dontrun{
  data("Bindea2013c_enrich")
  res <- Bindea2013c_enrich
  res <-  TimiPermFDR(resdata = res, geneset = geneset, gene = GP,
                       background = background, niter = 100, core = 20)
  head(score)
}
}
\author{
Chenyang Skylar Li
}
