% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{TimiPrePropress}
\alias{TimiPrePropress}
\title{Preprocess Transctiptomic profile}
\usage{
TimiPrePropress(marker, rna = NULL, cohort, log = TRUE)
}
\arguments{
\item{marker}{a vector of marker gene list.}

\item{rna}{a data.frame of transcriptomic profile, 
in which row names are genes and colnames are individuals.}

\item{cohort}{a vector of selected individuals.}

\item{log}{logical value: 
if TRUE, log transformation will be performed; The default is TRUE.}
}
\value{
Preprocessed transctiptomic profile
}
\description{
Given the transcriptomic profile, marker gene list, clinical information,
this function will return the marker gene expression in selected cohort,
after log transformation and  gene wise median normalization
}
\examples{
\dontrun{
  data("SKCM06info")
  data("SKCM06rna")
  data("Immune_Marker_n1326")
  dim(SKCM06info)
  info <- TimiCheckEvent(SKCM06info)
  dim(info)
  dim(SKCM06rna)
  rna <- TimiPrePropress(gene = Immune_Marker_n1326,rna = SKCM06rna,cohort = rownames(info))
  dim(rna)
}
}
\author{
Chenyang Skylar Li
}
