% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TNBCaPD1info}
\alias{TNBCaPD1info}
\title{GSE194040 TNBC chemoimmunotherapy cohort therapy response infomation}
\format{
A data frame with 26 rows and 1 variables:
\describe{
  \item{row}{Patient ID}
  \item{column}{Response: 1=responder, 0=non-responder }
}
}
\source{
Codes can be found in inst/extdata/build_TNBCaPD1_RNA_info.R
}
\usage{
data(TNBCaPD1info)
}
\description{
A dataset containing therapy response from 26 
Triple-negative breast cancer (TNBC) patients,
treated with the combination of anti-PD1 immunotherapy and chemotherpay
}
\references{
\url{"https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE194040"}
}
\keyword{datasets}
