% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TNBCaPD1rna}
\alias{TNBCaPD1rna}
\title{GSE194040 TNBC chemoimmunotherapy cohort microarray expression}
\format{
A data frame with 19,134 rows and 26 variables:
\describe{
  \item{row}{Gene}
  \item{column}{Sample(patient) ID}
}
}
\source{
Codes can be found in inst/extdata/build_TNBCaPD1_RNA_info.R
}
\usage{
data(TNBCaPD1rna)
}
\description{
A dataset containing 19,134 genes from 26 
Triple-negative breast cancer (TNBC) patients,
treated with the combination of anti-PD1 immunotherapy and chemotherpay
}
\references{
\url{"https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE194040"}
}
\keyword{datasets}
