% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChordDiagram.R
\name{TimiCellChord}
\alias{TimiCellChord}
\title{Chord diagram of Cell Interaction}
\usage{
TimiCellChord(
  resdata = NULL,
  select = NULL,
  dataset = NULL,
  group = NULL,
  color = NULL,
  condition = "Adjust.P.Value",
  cutoff = 0.05
)
}
\arguments{
\item{resdata}{TimiGP enrichment result generated from TimiEnrich}

\item{select}{A numeric vector of selected cell interactions according to "Index" column in resdata. 
Default selection is all inter-cell interactions(adjusted p value < 0.05).}

\item{dataset}{A value in one of 
c("Bindea2013","Bindea2013_Cancer","Charoentong2017", "Xu2018"
,"Newman2015", 
"Tirosh2016","Zheng2021","Other"). 
The first four options include default group and color settings. 
If you use other dataset or want to change group and colors, 
please choose "Other".}

\item{group}{A vector of self-defined cell groups, 
whose names are the all favorable and unfavorable cells types in selection.}

\item{color}{A vector of self-defined cell colors, 
whose names are the all favorable and unfavorable cells types in selection.}

\item{condition}{A value in one of 
c("P.Value","Adjust.P.Value","Permutation.FDR"),
which is the column name of resdata.
The default value is the "Adjust.P.Value".}

\item{cutoff}{A cutoff of condition used to filter cell interactions. 
The default cutoff is 0.05.}
}
\value{
Chord diagram of Cell Interaction
}
\description{
Chord diagram reveals the inter-cell interaction network.
The arrow points from favorable cell type A to unfavorable cell type B, 
which denotes that function of A greater than B associated with a good prognosis.
The width of the arrow represents -log10(Adjust.P.Value),namely,
the wider the arrow is, the smaller the adjusted p-value is.
}
\examples{
\dontrun{
  data("Bindea2013c_enrich")
  res <- Bindea2013c_enrich
  TimiCellChord(resdata = res,dataset = "Bindea2013_Cancer")
  TimiCellChord(resdata = res,dataset = "Bindea2013_Cancer",select = 1:10)
 
}
}
\author{
Chenyang Skylar Li
}
