% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network.R
\name{TimiCellNetwork}
\alias{TimiCellNetwork}
\title{Generate cell network files for Cytoscape}
\usage{
TimiCellNetwork(
  resdata = NULL,
  select = NULL,
  dataset = NULL,
  group = NULL,
  geneset = NULL,
  export = TRUE,
  path = NULL
)
}
\arguments{
\item{resdata}{TimiGP enrichment result generated from TimiEnrich}

\item{select}{a numeric vector of selected cell interactions according to "Index" column in resdata. 
Default selection is all functional interactions(adjusted p value < 0.05).}

\item{dataset}{a value in one of 
c("Bindea2013","Bindea2013_Cancer",
"Charoentong2017", "Xu2018",
"Newman2015", "Aran2017", "Luca2021", 
"Tirosh2016","Zheng2021","Other")
The first four options include default group and color settings. 
If you use other dataset or want to change group and colors, 
please choose "Other".}

\item{group}{a vector of self-defined cell groups, 
whose names are the all favorable and unfavorable cells types in selection.}

\item{geneset}{a data.frame of cell markers, 
in which the 1st is cell type, 
the 2nd column is the marker gene 
and the 3rd column is the name of the dataset(optional)}

\item{export}{a logical value. If TRUE, it will generate network files 
for Cytoscape analysis. The default value is TRUE.}

\item{path}{a directory to export the network files when "export = TRUE".}
}
\value{
A list of network required files(network,node,edge)
}
\description{
It generates three files that can be used to build network in Cytoscape:
1. network files: simple interaction file (cell_network.sif);
2. node attributes (cell_node.txt);
3. edge attributes (cell_edge.txt).
The function also returns a list of above files that can be modified in R.
}
\examples{
\dontrun{
  data("Bindea2013c_enrich")
  res <- Bindea2013c_enrich
  NET <- TimiCellNetwork(resdata = res,dataset = "Bindea2013_Cancer",export=T,path="./")
}
}
\author{
Chenyang Skylar Li
}
