% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dotplot.R
\name{TimiDotplot}
\alias{TimiDotplot}
\title{Dotplot}
\usage{
TimiDotplot(resdata = NULL, select = 1:5)
}
\arguments{
\item{resdata}{TimiGP enrichment result generated from TimiEnrich}

\item{select}{a numeric vector of selected cell pairs according to "Index" column in resdata. 
Default selection is top 5 enrichment}
}
\value{
Enrichment dotpot
}
\description{
A dotplot shows the results of selected cell pair enrichment.
The x-axis shows selected cell Pairs. The y-axis shows Enrichment Ratio.
The color of dots represents adjusted P-value. 
The size of dots represents the number of marker pairs 
shared by query pairs and annotation pairs.
}
\examples{
\dontrun{
  data("Bindea2013c_enrich")
  res <- Bindea2013c_enrich
  p1<-TimiDotplot(res)
  p1
  p2<-TimiDotplot(res,select = c(1:10))
  p2
 
}
}
\author{
Chenyang Skylar Li
}
