% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network.R
\name{TimiGeneNetwork}
\alias{TimiGeneNetwork}
\title{Generate gene network files for Cytoscape}
\usage{
TimiGeneNetwork(
  resdata = NULL,
  select = NULL,
  dataset = NULL,
  geneset = NULL,
  export = TRUE,
  path = NULL
)
}
\arguments{
\item{resdata}{TimiGP cox result generated from TimiCOX}

\item{select}{a character vector of selected gene pairs according to rownames in resdata. 
Default selection is all statistically significantly enriched cell pairs(adjusted p value < 0.05).}

\item{dataset}{a value in one of 
c("Galon2013_Cancer","Immune3", "Alex2020_Levi2019","Zhang2021S2","Other")
The first four options include default group and color settings. 
If you use other dataset or want to change group and colors, 
please choose "Other".}

\item{geneset}{a data.frame of cell markers, 
in which the 1st column is cell type, 
the 2nd column is the marker gene 
and the 3rd column is the name of the dataset(optional)}

\item{export}{a logical value. If TRUE, it will generate network files 
for Cytoscape analysis. The default value is TRUE.}

\item{path}{a directory to export the network files when "export = TRUE".}
}
\value{
A list of network required files(network,node,edge)
}
\description{
It generates three files that can be used to build network in Cytoscape:
1. network files: simple interaction file (network.sif);
2. node attributes (node.txt);
3. edge attributes (edge.txt).
The function also returns a list of above files that can be modified in R.
}
\examples{
\dontrun{
  data(Galon2013c_COX_MP_SKCM06)
  cox_res <- Galon2013c_COX_MP_SKCM06
  NET <- TimiGeneNetwork(resdata = cox_res,dataset = "Galon2013_Cancer")
}
}
\author{
Chenyang Skylar Li
}
