% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pairwise.R
\name{TimiGenePair}
\alias{TimiGenePair}
\title{Capture logical relation of marker pairs}
\usage{
TimiGenePair(rna = NULL)
}
\arguments{
\item{rna}{a data.frame of preprocessed transcriptomic profile}
}
\value{
a matrix of Marker Pair Score
}
\description{
Capture logical relation of any two marker pairs, 
and generate a matrix of Marker Pair Score:
1 or TRUE = the expression of gene A > that of gene B, 
0 or FALSE = the expression of gene A < that of gene B.
}
\examples{
\dontrun{
  data("SKCM06info")
  data("SKCM06rna")
  data("Immune_Marker_n1326")
  dim(SKCM06info)
  info <- TimiCheckEvent(SKCM06info)
  dim(info)
  dim(SKCM06rna)
  rna <- TimiPrePropress(gene = Immune_Marker_n1326,rna = SKCM06rna,cohort = rownames(info))
  # TimiGenePair
  mps <- TimiGenePair(rna)
  dim(mps)
}
}
\author{
Chenyang Skylar Li
}
